feat: Design review of Comments template and associated components
