--- packages/article-flag/src/style/index.android.js
@@ -1,13 +1,0 @@
-import { fontSizes } from "@times-components/styleguide";
-import sharedStyles from "./shared";
-
-const styles = {
-  ...sharedStyles,
-  title: {
-    ...sharedStyles.title,
-    fontSize: fontSizes.meta,
-    marginTop: 5
-  }
-};
-
-export default styles;
