+++ packages/article-flag/src/style/index.js
@@ -1,3 +1,13 @@
+import { fontSizes } from "@times-components/styleguide";
 import sharedStyles from "./shared";
 
-export default sharedStyles;
+const styles = {
+  ...sharedStyles,
+  title: {
+    ...sharedStyles.title,
+    fontSize: fontSizes.meta,
+    marginTop: 5
+  }
+};
+
+export default styles;
