+++ packages/article-flag/src/style/index.web.js
@@ -1,0 +1,3 @@
+import sharedStyles from "./shared";
+
+export default sharedStyles;
