+++ packages/article-main-comment/src/styles/index.web.js
@@ -1,8 +1,9 @@
 import { StyleSheet } from "react-native";
+import { spacing } from "@times-components/styleguide";
 import sharedStyles from "./shared";
 
 const styles = StyleSheet.create({
-  ...sharedStyles
+  ...sharedStyles,
 });
 
 export default styles;
