+++ packages/provider-test-tools/fixtures/article.json
@@ -1530,7 +1530,7 @@
     ]
   },
   "section": "business",
-  "shortHeadline": "A Test Short Headline",
+  "shortHeadline": "Let's keep A levels but scrap outdated GCSEs",
   "shortIdentifier": "37b27qd2s",
   "slug": "france-defies-may-over-russia",
   "standfirst": "How did one of Britain’s top young athletes end up running elite parties for swingers? Chris Reynolds Gordon tells Rick Broadbent about his extraordinary life",
